MM-19185 Fix tab for UnreadIndicator on Android
